Indev

+++ scripts/map/objects/IMapObject.cs
@@ -1,5 +1,5 @@
 public interface IMapObject
 {
 	int ObjectID {get;}
-	float Millisecond {get;}
+	int Millisecond {get;}
 }
